refactor to theme interface
