--- src/ui/ColorPalette.tsx
@@ -1,15 +1,0 @@
-export interface Theme {
-  lighter: string
-  light: string
-  dark: string
-  darker: string
-  accent: string
-}
-
-export const ColorPalette: Theme = {
-  lighter: '#fff',
-  light: '#f1f2f6',
-  dark: '#2f3542',
-  darker: '#1e272e',
-  accent: '#ff6348',
-}
